services/budgateway/gateway/tests/sdk/requirements.txt
openai>=1.0.0
anthropic>=0.31.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
httpx>=0.24.0
Pillow>=10.0.0
requests>=2.31.0